requirements.txt
seaborn==0.11.1
autograd==1.3
spatialmath_python==0.8.5
matplotlib==3.3.2
roboticstoolbox_python.egg==0.6.2
numpy==1.19.2
